+++ assets/scripts/TargetMatching.cs
@@ -960,6 +960,19 @@ public class TargetMatching : MonoBehaviour
         string fileName = GameObject.Find("SettingMenu").GetComponent<SettingsMenu>().getField((int)SettingsMenu.settings.gridSaveName);
 
         path += fileName;
+        if (System.IO.File.Exists(path))
+        {
+            bool val = PromptOverwrite();
+            if (val == false)
+            {
+                //Debug.Log("Quitting Save");
+                return;
+            }
+            else
+            {
+                // Debug.Log("Overwriting!!!");
+            }
+        }
 
         using (System.IO.StreamWriter file =
             new System.IO.StreamWriter(path, false))
@@ -992,7 +1005,7 @@ public class TargetMatching : MonoBehaviour
                 //CreateScalpHotSpot(p, r);
             }
 
-            GameObject.Find("SettingMenu").GetComponent<SettingsMenu>().incrementField((int)SettingsMenu.settings.gridSaveName);
+            //GameObject.Find("SettingMenu").GetComponent<SettingsMenu>().incrementField((int)SettingsMenu.settings.gridSaveName);
         }
     }
 
@@ -1029,6 +1042,9 @@ public class TargetMatching : MonoBehaviour
         usingGrid = true;
 
         GameObject.Find("Save Grids").GetComponent<Button>().interactable = true;
+        if (AskIfToSave()) {
+            ExportGrid(0);
+        }
     }
 
     public void ImportGrid()
@@ -1037,14 +1053,7 @@ public class TargetMatching : MonoBehaviour
         print(path);
         string fileName = GameObject.Find("SettingMenu").GetComponent<SettingsMenu>().getField((int)SettingsMenu.settings.gridLoadName);
 
-        //string[] fileNames = System.IO.Directory.GetFiles(path);
-
-
-        //System.IO.FileStream filestream = new System.IO.FileStream(fileNames[0],
-        //                                  System.IO.FileMode.Open,
-        //                                  System.IO.FileAccess.Read,
-        //                                  System.IO.FileShare.Read);
-        //System.IO.StreamReader file = new System.IO.StreamReader(filestream);
+       
 
         try {
             System.IO.FileStream filestream = new System.IO.FileStream(path + "/" + fileName,
@@ -1247,5 +1256,151 @@ public class TargetMatching : MonoBehaviour
     {
         if (initalized)
         logger.SetFileName(name);
+    }
+
+    bool PromptOverwrite()
+    {
+        using (var form1 = new System.Windows.Forms.Form())
+        {
+            System.Windows.Forms.Label text = new System.Windows.Forms.Label();
+            System.Windows.Forms.Button button1 = new System.Windows.Forms.Button();
+            System.Windows.Forms.Button button3 = new System.Windows.Forms.Button();
+            System.Windows.Forms.Button buttondefault = new System.Windows.Forms.Button();
+            buttondefault.Location = new System.Drawing.Point(-2000, -2000);
+
+            text.Text = text.Text = "A file exists at the Grid Save location specified! \nDo you want to overwrite?\n\nIf not: Cancel, then edit the Save Grid Field, \nthen Save Manually.";
+            text.Width = 1000;
+            text.Height = 70;
+            text.Location
+               = new System.Drawing.Point(10, 10);
+
+            // Set the text of button1 to "OK".
+            button3.Text = "Cancel Save";
+            // Set the position of the button on the form.
+            button3.Location = new System.Drawing.Point(text.Left, text.Height + text.Top + 10);
+            button3.BackColor = System.Drawing.Color.LightGreen;
+            button3.Width = 100;
+
+            // Set the text of button1 to "OK".
+            button1.Text = "Overwrite!";
+            // Set the position of the button on the form.
+            button1.Location = new System.Drawing.Point(button3.Left, button3.Height + button3.Top + 15);
+            button1.BackColor = System.Drawing.Color.LightYellow;
+            button1.Width = 100;
+            form1.Text = "CAUTION";
+            // Define the border style of the form to a dialog box.
+            form1.FormBorderStyle = System.Windows.Forms.FormBorderStyle.FixedDialog;
+            // Set the MaximizeBox to false to remove the maximize box.
+            form1.MaximizeBox = false;
+            // Set the MinimizeBox to false to remove the minimize box.
+            form1.MinimizeBox = false;
+            // Set the accept button of the form to button1.
+            form1.AcceptButton = button1;
+            form1.CancelButton = button3;
+            // Set the cancel button of the form to button2.
+            // Set the start position of the form to the center of the screen.
+            form1.StartPosition = System.Windows.Forms.FormStartPosition.CenterScreen;
+
+            form1.Height = 200;
+
+            button1.DialogResult = System.Windows.Forms.DialogResult.OK;
+            button3.DialogResult = System.Windows.Forms.DialogResult.Cancel;
+
+            //Add button1 to the form.
+            form1.Controls.Add(buttondefault);
+            form1.Controls.Add(button1);
+            //Add button2 to the form.
+            form1.Controls.Add(button3);
+
+            form1.Controls.Add(text);
+            System.Windows.Forms.DialogResult retval = form1.ShowDialog();
+            // Display the form as a modal dialog box.
+            if (retval == System.Windows.Forms.DialogResult.Cancel)
+            {
+                //Debug.Log("Canceled save");
+                return false;
+            }
+            if (retval == System.Windows.Forms.DialogResult.OK)
+            {
+                //Debug.Log("accepted");
+                return true;
+            }
+
+
+        }
+        return false;
+    }
+
+    bool AskIfToSave()
+    {
+        using (var form1 = new System.Windows.Forms.Form())
+        {
+            System.Windows.Forms.Label text = new System.Windows.Forms.Label();
+            System.Windows.Forms.Button button1 = new System.Windows.Forms.Button();
+            System.Windows.Forms.Button button3 = new System.Windows.Forms.Button();
+            System.Windows.Forms.Button buttondefault = new System.Windows.Forms.Button();
+            buttondefault.Location = new System.Drawing.Point(-2000, -2000);
+
+            text.Text = "Successfully calibrated Grid! Do you want to save to \nGrid Save location?";
+            text.Width = 1000;
+            text.Height = 50;
+            text.Location
+               = new System.Drawing.Point(10, 10);
+
+            // Set the text of button1 to "OK".
+            button3.Text = "Yes";
+            // Set the position of the button on the form.
+            button3.Location = new System.Drawing.Point(text.Left, text.Height + text.Top + 10);
+            button3.BackColor = System.Drawing.Color.LightGreen;
+            button3.Width = 100;
+
+            // Set the text of button1 to "OK".
+            button1.Text = "No";
+            // Set the position of the button on the form.
+            button1.Location = new System.Drawing.Point(button3.Left, button3.Height + button3.Top + 15);
+            button1.BackColor = System.Drawing.Color.Pink;
+            button1.Width = 100;
+            form1.Text = "";
+            // Define the border style of the form to a dialog box.
+            form1.FormBorderStyle = System.Windows.Forms.FormBorderStyle.FixedDialog;
+            // Set the MaximizeBox to false to remove the maximize box.
+            form1.MaximizeBox = false;
+            // Set the MinimizeBox to false to remove the minimize box.
+            form1.MinimizeBox = false;
+            // Set the accept button of the form to button1.
+            form1.AcceptButton = button1;
+            form1.CancelButton = button3;
+            // Set the cancel button of the form to button2.
+            // Set the start position of the form to the center of the screen.
+            form1.StartPosition = System.Windows.Forms.FormStartPosition.CenterScreen;
+
+            form1.Height = 200;
+
+            button1.DialogResult = System.Windows.Forms.DialogResult.Cancel;
+            button3.DialogResult = System.Windows.Forms.DialogResult.OK;
+
+            //Add button1 to the form.
+            form1.Controls.Add(buttondefault);
+            form1.Controls.Add(button1);
+            //Add button2 to the form.
+            form1.Controls.Add(button3);
+
+            form1.Controls.Add(text);
+            System.Windows.Forms.DialogResult retval = form1.ShowDialog();
+            // Display the form as a modal dialog box.
+            if (retval == System.Windows.Forms.DialogResult.Cancel)
+            {
+                //Debug.Log("Canceled save");
+                return false;
+            }
+            if (retval == System.Windows.Forms.DialogResult.OK)
+            {
+                //Debug.Log("accepted");
+                return true;
+            }
+
+
+        }
+        return false;
     }
 }
